📦 refactor(index.js): move index.js from lib/ to src/
