--- lib/index.js
@@ -1,13 +1,0 @@
-const { createEventHandler } = require("../src/event-handler");
-const { createMiddleware } = require("../src/middleware");
-const { createWebhooksApi } = require("../src");
-const { sign } = require("../src/sign");
-const { verify } = require("../src/verify");
-
-module.exports = {
-	createEventHandler,
-	createMiddleware,
-	createWebhooksApi,
-	sign,
-	verify
-}
