+++ test/integration/smoke-test.js
@@ -3,7 +3,7 @@ const { mock } = require("simple-mock");
 
 test("@octokit/webhooks", (t) => {
   const emitWarningMock = mock(process, "emitWarning");
-  const { createWebhooksApi } = require("../../lib");
+  const { createWebhooksApi } = require("../../src");
   const api = createWebhooksApi({ secret: "mysecret" });
 
   t.type(api.sign, "function");
@@ -22,7 +22,7 @@ test('require("@octokit/webhooks").sign', (t) => {
   const emitWarningMock = mock(process, "emitWarning");
 
   t.doesNotThrow(() => {
-    const { sign } = require("../../lib");
+    const { sign } = require("../../src");
     sign("1234", {});
   });
   t.false(emitWarningMock.called);
@@ -34,7 +34,7 @@ test('require("@octokit/webhooks").verify', (t) => {
   const emitWarningMock = mock(process, "emitWarning");
 
   t.doesNotThrow(() => {
-    const { verify } = require("../../lib");
+    const { verify } = require("../../src");
     verify("1234", {}, "randomSignature");
   });
   t.false(emitWarningMock.called);
@@ -46,7 +46,7 @@ test('require("@octokit/webhooks").createEventHandler', (t) => {
   const emitWarningMock = mock(process, "emitWarning");
 
   t.doesNotThrow(() => {
-    const { createEventHandler } = require("../../lib");
+    const { createEventHandler } = require("../../src");
     createEventHandler();
   });
   t.false(emitWarningMock.called);
@@ -58,7 +58,7 @@ test('require("@octokit/webhooks).createMiddleware")', (t) => {
   const emitWarningMock = mock(process, "emitWarning");
 
   t.doesNotThrow(() => {
-    const { createMiddleware } = require("../../lib");
+    const { createMiddleware } = require("../../src");
     createMiddleware({ secret: "1234" });
   });
   t.false(emitWarningMock.called);
