+++ test/unit/middleware-constructor-test.js
@@ -1,6 +1,6 @@
 const test = require("tap").test;
 
-const { createMiddleware } = require("../../lib");
+const { createMiddleware } = require("../../src");
 
 test("options: none", (t) => {
   t.throws(createMiddleware);
